[CLEANUP] Modified function syntax for consistency [CLEANUP] Move registerServiceWorker back to /src [MISC] Lengthened timer for spinning logo
