+++ src/stylesheets/index.css
@@ -12,7 +12,7 @@ ol, ul {
 }
 
 .Header-logo {
-  animation: Header-logo-spin infinite 20s linear;
+  animation: Header-logo-spin infinite 30s linear;
   height: 80px;
 }
